+++ Binaries/dcmtk-source/dcmimgle/dimo2img.cc
@@ -94,8 +94,8 @@ DiMono2Image::DiMono2Image(const DiColorImage *image,
 
 
 DiMono2Image::DiMono2Image(const DiMonoImage *image,
-                           const signed long left_pos,
-                           const signed long top_pos,
+                           const signed int left_pos,
+                           const signed int top_pos,
                            const Uint16 src_cols,
                            const Uint16 src_rows,
                            const Uint16 dest_cols,
@@ -171,8 +171,8 @@ DiImage *DiMono2Image::createImage(const unsigned int fstart,
 }
 
 
-DiImage *DiMono2Image::createScale(const signed long left_pos,
-                                   const signed long top_pos,
+DiImage *DiMono2Image::createScale(const signed int left_pos,
+                                   const signed int top_pos,
                                    const unsigned int src_cols,
                                    const unsigned int src_rows,
                                    const unsigned int dest_cols,
